Implement rotor middleware
Now instead of config.host and config.port, we have config.hosts, with a
config.host= convenience method to use a single host. If more than one
host is specified, use the Rotor middleware, which will round-robin
between the hosts. In the case of a connection failure, Rotor will try
the next host until all are exhausted, at which point it will raise a
NoServerError.