Deal with parents of child documents when loading from direct search
If you perform a search directly on a child document class, there's no
parent collection from which to populate the parent. So in that case we
should materialize the parent by performing a get operation on the ID
returned in the hit. Also if we re-save the document, hold on to the
parent to pass it back in the update call.