Fail fast when attempting to double-create document in bulk operation
It's not possible to call #save twice on the same transient document
within a single bulk operation. The reason is that the second #save
should initiate an update; however, in order to perform an update, we
need the document's ID. We don't get the ID back until we get a response
from the bulk operation.