Expose raw hits when enumerating over search results
The #find_each and #find_in_batches enumerators now yield a second
argument, which is a Mash of the raw hit corresponding to the result.
This gives access to hit metadata (e.g. highlighting). The #each method
no longer aliases to #find_each; instead, it enumerates over only the
result documents, without the hits (so other Enumerable methods work in
an intuitive way).