Remove #enumerate_each and fix behavior of ChildCollectionProxy#each
Returned the #enumerate_each method in favor of constructing a
self-referential enumerator if no block is given to #find_each (as we do
elsewhere). Also made more sensible behavior for the
ChildCollectionProxy -- it now overrides the #each method, yielding the
transient children along with the persisted results, but doesn't touch
the #find_each etc. methods, which semantically should only be yielding
results from the underlying store.