Change Elastictastic::Resource to Elastictastic::NestedDocument
We are going to need functionality on nested documents that doesn't make
sense for top-level documents, so it's no longer appropriate to have
Elastictstic::Document "inherit" from Elastictastic::Resource. Instead,
Document and NestedDocuments both mix in the modules they need. Most of
the methods on Elastictastic::Resource now live in
Elastictastic::Properties, which is mixed in to both Document and
NestedDocument.