Support :auto_flush option for bulk persistence
If this option is passed in, the bulk persister will flush itself after
the specified number of operations is buffered. Good for very large bulk
operations.

Also got rid of CancelBulkOperation; this implies bulk operations are
like transactions, which they're not. Also, use of this error in an
auto-flushing bulk operation would essentially cause nondeterministic
behavior.