Only perform last operation on each document in bulk block
In the case where we perform multiple operations on the same document in
a given bulk block, we only want to pass the last operation. Otherwise,
the first operation will cause the version to increment during the
course of the bulk flush, and subsequent operations will have version
conflict errors. Since ElasticSearch does not support partial write
operations, simply performing the last operation on the document will
have the same effect as performing all of them in sequence.