Clear bulk buffer before processing response
In the case where the bulk strategy has an auto-flush, we may have a
handler for concurrency problems that performs a new write operation.
Previously, this write operation would be added to the bulk buffer; when
it is added, the bulk buffer sees that it is over its auto-flush
threshold and attempts to re-flush the entire buffer, including all the
previously buffered operations. So, clear the buffer before processing
the response.