Don't throw an error on out-of-bounds pages
If we request a "from" that's out of range of the total hits, ES will
return an empty collection of hits. That behavior should be passed
through to the client. The previous implementation of multi-search had a
too-clever-by-half method of determining whether to memoize a hits
collection: if the hits are non-empty, or the total is 0, memoize it.
If not, we were assuming the results were from a count query. However,
an out-of-bounds result would look the same as a count query, but in
this case, we do want to memoize the empty result set. So, always just
make it explicit whether we're populating the results of a count or a
query_then_fetch.

As part of this refactor, MultiSearch objects now let you mix-and-match
scopes to perform a count on and scopes to perform a full query on. The
old API is retained and the mix-and-match capability is untested and
thus unofficial.