Remove Faraday dependency
It's been a bit of a pain for a while, but in particular I would like:

* Persistent connections with Excon
* Timeouts with Excon

Neither of which Faraday does right now. Created a simple adapter
framework that also supports "middlewares" via composition. Currently
just support Net::HTTP and Excon, but other libraries will be trivial
to add support for.