Support multiple ID arguments to ::create_or_update and ::update
For ::create_or_update, it's just a convenience, as any needed updates
will still be executed one at a time. For ::update, though, a multiget
will automatically be used to load the documents initially, which should
give a performance boost when updating a lot of documents at once by
avoiding N GET requests.